+++ test/package.json
@@ -5,6 +5,7 @@
   "proxy": "http://localhost:9000/",
   "devDependencies": {
     "body-parser": "^1.16.1",
+    "cors": "^2.8.1",
     "express": "^4.14.1",
     "express-cors": "0.0.3",
     "path": "^0.12.7",
