+++ test/src/controller.js
@@ -1,5 +1,5 @@
 exports.helloworld = function(req, res, next) {
   res.status(200).json({
-    message: "Hello world!"
+    message: 'hello'
   })
 }
