--- test/src/index.pug
@@ -1,5 +1,0 @@
-html
-  head
-    title= title
-  body
-    h1= message
