+++ test/src/server.js
@@ -2,8 +2,10 @@ const express = require('express');
 const request = require('request');
 const bodyParser = require('body-parser');
 const path = require('path');
-const cors = require('express-cors')
+const cors = require('cors')
 const app = express()
+const router = require('./router');
+
 
 const bonfig = {
   clientId: "uY-l55ombZgi1T9IF1Jl5Cb3wGZqw9uC444WRPHPK6TOu6aIFELNvtIZA3HWqngr",
@@ -12,7 +14,8 @@ const bonfig = {
  clientSecret: process.env.GENIUS_CLIENT_SECRET
 }
 
-// app.use(cors())
+app.use(cors())
+router(app)
 app.use(bodyParser.urlencoded({ extended: true }));
 app.use(bodyParser.json())
 
@@ -27,48 +30,48 @@ app.get('/', (request, response, next) => {
   response.redirect(authUrl)
 })
 
-app.get('/test', (req, res, next) => {
-  let ops = {
-    url: 'https://api.genius.com/artists/16775/songs',
-    headers: {
-      Authorization: 'Bearer j4DQ4ILmQIj07lZA6P_j_2ZjTrG_db2Bxg2aIvLN7tVaq0UxgSgqh8He1T3o28UM',
-      Accept: 'application/json',
-    }
-  }
-
-  let options = {
-    url: 'https://api.genius.com/oauth/token',
-    form: {
-      code: req.query.code,
-      client_secret: process.env.GENIUS_CLIENT_SECRET,
-      grant_type: 'authorization_code',
-      client_id: bonfig.clientId,
-      redirect_uri: bonfig.redirectUri,
-      response_type: 'code'
-    }
-  }
-
-  request.post(options, (error, response) => {
-    console.log('status code:', response.statusCode)
-    if (response.statusCode > 399) {
-      console.log('error', error)
-    } else {
-      console.log(response.statusCode)
-      let body = JSON.parse(response.body)
-      console.log(body.access_token)
-    }
-  })
-
-
-app.use(express.static(path.resolve(__dirname, '..', 'build')));
-
-  request.get(ops, (error, response) => {
-    console.log(error)
-    let body = JSON.parse(response.body)
-    console.log(body.response.songs)
-  })
-  res.sendFile(path.join(__dirname, '..', '/build', 'index.html'));
-})
+// app.get('/test', (req, res, next) => {
+//   let ops = {
+//     url: 'https://api.genius.com/artists/16775/songs',
+//     headers: {
+//       Authorization: 'Bearer j4DQ4ILmQIj07lZA6P_j_2ZjTrG_db2Bxg2aIvLN7tVaq0UxgSgqh8He1T3o28UM',
+//       Accept: 'application/json',
+//     }
+//   }
+//
+//   let options = {
+//     url: 'https://api.genius.com/oauth/token',
+//     form: {
+//       code: req.query.code,
+//       client_secret: process.env.GENIUS_CLIENT_SECRET,
+//       grant_type: 'authorization_code',
+//       client_id: bonfig.clientId,
+//       redirect_uri: bonfig.redirectUri,
+//       response_type: 'code'
+//     }
+//   }
+//
+//   request.post(options, (error, response) => {
+//     console.log('status code:', response.statusCode)
+//     if (response.statusCode > 399) {
+//       console.log('error', error)
+//     } else {
+//       console.log(response.statusCode)
+//       let body = JSON.parse(response.body)
+//       console.log(body.access_token)
+//     }
+//   })
+//
+//
+// app.use(express.static(path.resolve(__dirname, '..', 'build')));
+//
+//   request.get(ops, (error, response) => {
+//     console.log(error)
+//     let body = JSON.parse(response.body)
+//     console.log(body.response.songs)
+//   })
+//   res.sendFile(path.join(__dirname, '..', '/build', 'index.html'));
+// })
 
 app.listen(9000, () => {
   console.log('go to http://localhost:9000/ and authenticate. Your access token will then appear in this console.')
